refactor: move cardview to separate folder and update imports
